WLA-430: Resolving configuration
