+++ src/main/java/uk/gov/hmcts/reform/sscs/personalisation/SyaAppealCreatedAndReceivedPersonalisation.java
@@ -1,23 +1,23 @@
 package uk.gov.hmcts.reform.sscs.personalisation;
 
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.ANYTHING;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.DATE_OF_BIRTH;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.DATE_OF_MRN;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.DISABLED_ACCESS;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.HAVE_AN_APPOINTEE;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.HAVE_A_REPRESENTATIVE;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.HEARING_LOOP;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.LANGUAGE_INTERPRETER;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.NINO;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.ORGANISATION;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.OTHER_ARRANGEMENTS;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.REASON_FOR_LATE_APPEAL;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.REASON_FOR_NO_MRN;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.RECEIVE_TEXT_MESSAGE_REMINDER;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.SIGN_INTERPRETER;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.WHAT_DISAGREE_WITH;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.WHY_DISAGREE_WITH;
-import static uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey.getPersonalisationKey;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.ANYTHING;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.DATE_OF_BIRTH;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.DATE_OF_MRN;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.DISABLED_ACCESS;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.HAVE_AN_APPOINTEE;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.HAVE_A_REPRESENTATIVE;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.HEARING_LOOP;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.LANGUAGE_INTERPRETER;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.NINO;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.ORGANISATION;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.OTHER_ARRANGEMENTS;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.REASON_FOR_LATE_APPEAL;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.REASON_FOR_NO_MRN;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.RECEIVE_TEXT_MESSAGE_REMINDER;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.SIGN_INTERPRETER;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.WHAT_DISAGREE_WITH;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.WHY_DISAGREE_WITH;
+import static uk.gov.hmcts.reform.sscs.config.PersonalisationKey.getPersonalisationKey;
 import static uk.gov.hmcts.reform.sscs.service.NotificationUtils.hasAppointee;
 
 import java.time.LocalDate;
@@ -47,7 +47,7 @@ import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Subscription;
 import uk.gov.hmcts.reform.sscs.config.AppConstants;
 import uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration;
-import uk.gov.hmcts.reform.sscs.config.PersonalisationConfiguration.PersonalisationKey;
+import uk.gov.hmcts.reform.sscs.config.PersonalisationKey;
 import uk.gov.hmcts.reform.sscs.domain.SscsCaseDataWrapper;
 import uk.gov.hmcts.reform.sscs.domain.SubscriptionWithType;
 import uk.gov.hmcts.reform.sscs.service.conversion.LocalDateToWelshStringConverter;
