+++ src/main/resources/application.yaml
@@ -855,67 +855,67 @@ amqp:
 
 personalisation:
   english:
-    ATTENDING_HEARING: "Attending the hearing: "
-    YES: "yes"
-    NO: "no"
-    DATES_NOT_ATTENDING: "Dates you can't attend: "
-    DATE_OF_MRN: "Date of MRN: "
-    REASON_FOR_LATE_APPEAL: "Reason for late appeal: "
-    REASON_FOR_NO_MRN: "Reason for no MRN: "
-    NAME: "Name: "
-    DATE_OF_BIRTH: "Date of birth: "
-    NINO: "National Insurance number: "
-    ADDRESS: "Address: "
-    EMAIL: "Email: "
-    PHONE: "Phone: "
-    RECEIVE_TEXT_MESSAGE_REMINDER: "Receive text message reminders: "
-    MOBILE: "Mobile number: "
-    HAVE_AN_APPOINTEE: "Have an appointee: "
-    NOT_PROVIDED: "Not provided"
-    HAVE_A_REPRESENTATIVE: "Have a representative: "
-    ORGANISATION: "Organisation: "
-    WHAT_DISAGREE_WITH: "What you disagree with: "
-    WHY_DISAGREE_WITH: "Why you disagree with it: "
-    ANYTHING: "Anything else you want to tell the tribunal: "
-    LANGUAGE_INTERPRETER: "Language interpreter: "
-    SIGN_INTERPRETER: "Sign interpreter: "
-    HEARING_LOOP: "Hearing loop: "
-    DISABLED_ACCESS: "Disabled access: "
-    OTHER_ARRANGEMENTS: "Any other arrangements: "
-    REQUIRED: "Required"
-    NOT_REQUIRED: "Not required"
-    OTHER: "Other"
+    attendingHearing: "Attending the hearing: "
+    yes: "yes"
+    no: "no"
+    datesNotAtteding: "Dates you can't attend: "
+    dateOfMrn: "Date of MRN: "
+    reasonForLateAppeal: "Reason for late appeal: "
+    reasonForNoMrn: "Reason for no MRN: "
+    name: "Name: "
+    dateOfBirth: "Date of birth: "
+    nino: "National Insurance number: "
+    address: "Address: "
+    email: "Email: "
+    phone: "Phone: "
+    receiveTextMessageReminder: "Receive text message reminders: "
+    mobile: "Mobile number: "
+    haveAnAppointee: "Have an appointee: "
+    notProvided: "Not provided"
+    haveARepresentative: "Have a representative: "
+    organisation: "Organisation: "
+    whatDisagreeWith: "What you disagree with: "
+    whyDisagreeWith: "Why you disagree with it: "
+    anything: "Anything else you want to tell the tribunal: "
+    languageInterpreter: "Language interpreter: "
+    signInterpreter: "Sign interpreter: "
+    hearingLoop: "Hearing loop: "
+    disabledAccess: "Disabled access: "
+    otherArrangements: "Any other arrangements: "
+    required: "Required"
+    notRequired: "Not required"
+    other: "Other"
   welsh:
-    ATTENDING_HEARING: "Ydych chi'n bwriadu mynychu'r gwrandawiad: "
-    YES: "ydw"
-    NO: "nac ydw"
-    DATES_NOT_ATTENDING: "Dyddiadau na allwch fynychu: "
-    DATE_OF_MRN: "Dyddiad yr MRN: "
-    REASON_FOR_LATE_APPEAL: "Rheswm dros apêl hwyr: "
-    REASON_FOR_NO_MRN: "Rheswm dros ddim MRN: "
-    NAME: "Enw: "
-    DATE_OF_BIRTH: "Dyddiad geni: "
-    NINO: "Rhif Yswiriant Gwladol: "
-    ADDRESS: "Cyfeiriad: "
-    EMAIL: "E-bost: "
-    PHONE: "Rhif ffôn: "
-    RECEIVE_TEXT_MESSAGE_REMINDER: "Eisiau negeseuon testun atgoffa: "
-    MOBILE: "Rhif ffôn symudol: "
-    HAVE_AN_APPOINTEE: "Have an appointee: "
-    NOT_PROVIDED: "Nis ddarparwyd"
-    HAVE_A_REPRESENTATIVE: "A oes gennych chi gynrychiolydd: "
-    ORGANISATION: "Sefydliad: "
-    WHAT_DISAGREE_WITH: "Beth ydych chi’n anghytuno ag o: "
-    WHY_DISAGREE_WITH: "Pam ydych chi’n anghytuno ag o: "
-    ANYTHING: "Unrhyw beth arall yr hoffech ddweud wrth y tribiwnlys: "
-    LANGUAGE_INTERPRETER: "Dehonglydd iaith arwyddion: "
-    SIGN_INTERPRETER: "Dehonglydd iaith arwyddion: "
-    HEARING_LOOP: "Dolen glyw: "
-    DISABLED_ACCESS: "Mynediad i bobl anab: "
-    OTHER_ARRANGEMENTS: "Unrhyw drefniadau eraill: "
-    REQUIRED: "Gofynnol"
-    NOT_REQUIRED: "Dim yn ofynnol"
-    OTHER: "Other"
+    attendingHearing: "Ydych chi'n bwriadu mynychu'r gwrandawiad: "
+    yes: "ydw"
+    no: "nac ydw"
+    datesNotAtteding: "Dyddiadau na allwch fynychu: "
+    dateOfMrn: "Dyddiad yr MRN: "
+    reasonForLateAppeal: "Rheswm dros apêl hwyr: "
+    reasonForNoMrn: "Rheswm dros ddim MRN: "
+    name: "Enw: "
+    dateOfBirth: "Dyddiad geni: "
+    nino: "Rhif Yswiriant Gwladol: "
+    address: "Cyfeiriad: "
+    email: "E-bost: "
+    phone: "Rhif ffôn: "
+    receiveTextMessageReminder: "Eisiau negeseuon testun atgoffa: "
+    mobile: "Rhif ffôn symudol: "
+    haveAnAppointee: "Have an appointee: "
+    notProvided: "Nis ddarparwyd"
+    haveARepresentative: "A oes gennych chi gynrychiolydd: "
+    organisation: "Sefydliad: "
+    whatDisagreeWith: "Beth ydych chi’n anghytuno ag o: "
+    whyDisagreeWith: "Pam ydych chi’n anghytuno ag o: "
+    anything: "Unrhyw beth arall yr hoffech ddweud wrth y tribiwnlys: "
+    LanguageInterpreter: "Dehonglydd iaith arwyddion: "
+    signInterpreter: "Dehonglydd iaith arwyddion: "
+    hearingLoop: "Dolen glyw: "
+    disabledAccess: "Mynediad i bobl anab: "
+    otherArrangements: "Unrhyw drefniadau eraill: "
+    required: "Gofynnol"
+    notRequired: "Dim yn ofynnol"
+    other: "Other"
 
 
 ---
